Naming conventions like fuel_part.go instead fuelpart.go, refactor of tank.go and fuel_part.go instead model.go, added health check ping, added gomod, new Dockerfile, updated scripts start-app.sh, test.sh
