+++ microservice/challenge-3/booster/app/handler/ping.go
@@ -1,0 +1,10 @@
+package handler
+
+import (
+	"github.com/b-yond-infinite-network/amaze-us/microservice/challenge-3/booster/app/model"
+	"net/http"
+)
+
+func Ping(w http.ResponseWriter) {
+	respondJSON(w, http.StatusOK, model.GetPing())
+}
