+++ microservice/challenge-3/booster/app/model/ping.go
@@ -1,0 +1,11 @@
+package model
+
+type Ping struct {
+	Ping	string `json:"ping"`
+}
+
+const pong = "pong"
+
+func GetPing() *Ping {
+	return &Ping{Ping: pong}
+}
