+++ microservice/challenge-3/booster/app/model/tank.go
@@ -1,8 +1,6 @@
 package model
 
 import (
-	"time"
-
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
 )
@@ -20,23 +18,6 @@ func (p *Tank) Archive() {
 
 func (p *Tank) Restore() {
 	p.Archived = false
-}
-
-type FuelPart struct {
-	gorm.Model
-	Title     string     `json:"title"`
-	Priority  string     `gorm:"type:ENUM('0', '1', '2', '3');default:'0'" json:"priority"`
-	Deadline  *time.Time `gorm:"default:null" json:"deadline"`
-	Done      bool       `json:"done"`
-	TankID    uint       `json:"tank_id"`
-}
-
-func (t *FuelPart) Complete() {
-	t.Done = true
-}
-
-func (t *FuelPart) Undo() {
-	t.Done = false
 }
 
 // DBMigrate will create and migrate the tables, and then make the some relationships if necessary
