+++ microservice/challenge-3/booster/go.mod
@@ -1,0 +1,11 @@
+module github.com/b-yond-infinite-network/amaze-us/microservice/challenge-3/booster
+
+go 1.14
+
+require (
+	github.com/DATA-DOG/go-sqlmock v1.5.0
+	github.com/gorilla/mux v1.8.0
+	github.com/jinzhu/gorm v1.9.16
+	github.com/onsi/ginkgo v1.14.2
+	github.com/onsi/gomega v1.10.4
+)
