+++ microservice/challenge-3/booster/start-app.sh
