+++ microservice/challenge-3/booster/test.sh
@@ -1,1 +1,1 @@
-go test --cover -v ./tests/...
+go test --cover -v ./...
